fix: passing options correctly
