+++ package/src/index.tsx
@@ -14,7 +14,6 @@ export * from './hooks/useAssetAnimator'
 export * from './hooks/useConfigureAssetShadow'
 export * from './hooks/useEntityInScene'
 export * from './hooks/useLightEntity'
-export * from './hooks/useAmbientOcclusionOptions'
 export * from './hooks/useResource'
 
 // utilities
